Refactor: split githubSync, then add ETag conditional reads (#77, #69) (#107)
* refactor(githubSync): split 1602-line module into pull/push/classify (#77)

`src/utils/githubSync.ts` had grown to ~1602 lines and folded the pull
orchestration, push orchestration, the PullClassification kinds, the
zipball fast path, and every shared helper into one file. Split it
into focused submodules under `src/utils/githubSync/`:

- `internal.ts`     — path computation (notePath/pushPath/
  collectLocalFolderRepoPaths), MIME guessing, note serialization
  + parser (serializeNote/normalizeForPush/parseNote/parseInlineArrayField),
  the canonical-form equality helper, and the maybeEncryptForPush /
  maybeDecryptFromPull vault-key wrappers. Shared by both halves.
- `syncClassify.ts` — the `PullClassification` discriminated union and
  the `PullOutcome` shape. Was a ~250-line type block embedded in the
  pull file; it now lives on its own so the kinds (the six core +
  attachment/folder/settings extensions) can be referenced and
  extended without touching the orchestration files.
- `syncPull.ts`     — `pullFromGitHub`, `pullFromZipball`, and the
  zipball attachment side-channel cache (`takeZipballAttachmentBytes`).
- `syncPush.ts`     — `syncToGitHub`, the in-memory uploaded-blob-SHA
  cache, and the `_resetUploadedShaCache` test hook.

The top-level `src/utils/githubSync.ts` is now a thin barrel that
re-exports the public surface — every external import path
(`@/utils/githubSync` / `../utils/githubSync` / `./githubSync`) keeps
working untouched. No test or downstream import was updated.

Behavior-preserving: 2378/2378 jest tests pass, `npm run lint` clean,
`npx tsc --noEmit` clean.


* perf(githubSync): cache ETags on pull-side blob/tree reads (#69)

GitHub's authenticated REST budget is 5 000 req/hr, but a busy noteser
session re-reads the same recursive tree + blob SHAs every sync. Each
of those reads is a payload + a quota hit even when the content is
unchanged.

Add a two-tier ETag-conditional cache the pull uses on every blob and
tree read:
- `src/utils/githubETagCache.ts` — keyed by `(repo, sha)`; in-memory
  Map for the hot tier, idb-keyval for the cold tier so the next
  session starts warm. Cold misses delegate to the bare
  `getBlobContent` / `getTreeMap` so existing test mocks of
  `../utils/github` still intercept the read. Once an ETag is in cache,
  follow-up reads send `If-None-Match` directly via `githubFetch`; a
  304 short-circuits to the cached content with no body decode, a 200
  refreshes the entry.
- `src/utils/githubSync/syncPull.ts` — swap the four bare-helper call
  sites (the recursive tree fetch + remote-body / ancestor /
  vault-settings / .gitignore blob fetches) to the conditional
  variants. The recursive tree fetch already used `?recursive=1`, so
  the issue's "batch where the API allows" requirement is met.

Push path stays as-is (`syncPush.ts` still imports the bare helpers
from `../github`), so push semantics are byte-identical to pre-PR.

Tests (`src/__tests__/githubETagCache.test.ts`, 14 cases) cover:
- cold-cache miss → bare helper delegation, no conditional fetch
- pre-seeded ETag + 304 reply → cached content returned, bare helper
  NOT called, `If-None-Match: <etag>` sent
- pre-seeded ETag + 200 reply → new content returned, cache refreshed,
  next 304 read reflects the new ETag
- per-(repo, sha) keying: different sha / different repo bypass the
  cache cleanly
- invalidate hooks force the next read to delegate again

Verification: `npm run lint` clean, `npx tsc --noEmit` clean, jest
2388/2388 green (192 suites, +10 from Phase 1 net).


---------
