+++ src/utils/githubSync.ts
@@ -1,1602 +1,52 @@
-import type JSZip from 'jszip'
-import type { Note, Folder, SyncRepo } from '@/types'
-// Import from the dedicated sanitiser module, NOT from './export'. The
-// latter statically imports jszip + file-saver at module scope, and
-// githubSync sits on the app-init path (useGitHubSync → useAutoSync), so
-// importing sanitizeFilename from './export' dragged jszip + file-saver
-// into the first-load bundle. './sanitizeFilename' is jszip-free.
-import { sanitizeFilename } from './sanitizeFilename'
-import {
-  getBranchRefSha,
-  getCommitTreeSha,
-  getTreeMap,
-  createBlob,
-  createBlobBinary,
-  createTree,
-  createCommit,
-  updateBranchRef,
-  getBlobContent,
-  getBlobBytes,
-  gitBlobSha,
-  gitBlobShaBytes,
-  fetchZipball,
-  type GitTreeEntry,
-} from './github'
-import { threeWayMerge } from './lineDiff'
-import {
-  isAttachmentPath,
-  listAttachmentPaths,
-  getAttachmentBlob,
-  getAttachmentGitSha,
-  getAttachmentTombstones,
-  clearAttachmentTombstones,
-} from './attachments'
-import { encryptNoteContent, decryptNoteContent, isEncryptedContent } from './vaultCrypto'
-import { getVaultKey, VaultLockedError } from './vaultKey'
-
-// Encrypt note body for push, if the vault is unlocked AND encryption
-// is enabled at the settings layer. Returns the original content when
-// encryption is off — keeps the call sites tidy and ensures push works
-// in the default (unencrypted) configuration.
-async function maybeEncryptForPush(content: string): Promise<string> {
-  let enabled = false
-  try {
-    const { useSettingsStore } = await import('@/stores/settingsStore')
-    enabled = useSettingsStore.getState().vaultEncryptionEnabled
-  } catch {
-    // Test envs without the store — treat as disabled.
-  }
-  if (!enabled) return content
-  const key = getVaultKey()
-  if (!key) {
-    // Encryption is on but the user hasn't unlocked. Bail with a typed
-    // error the UI can catch to prompt for the passphrase.
-    throw new VaultLockedError('Push aborted — vault is encrypted but locked. Unlock to continue.')
-  }
-  return await encryptNoteContent(content, key)
-}
-
-// Decrypt remote note content on pull. Pass-through when the content
-// isn't an encrypted envelope. Throws VaultLockedError when an
-// envelope is present but the user hasn't unlocked yet — caller catches.
-async function maybeDecryptFromPull(content: string): Promise<string> {
-  if (!isEncryptedContent(content)) return content
-  const key = getVaultKey()
-  if (!key) {
-    throw new VaultLockedError('Pull skipped — remote blob is encrypted but vault is locked.')
-  }
-  return await decryptNoteContent(content, key)
-}
-
-export interface SyncResult {
-  unchanged: boolean
-  created: number
-  updated: number
-  deleted: number
-  commitSha: string
-  commitUrl: string | null
-}
-
-// ── Path computation ────────────────────────────────────────────────────────
-// Mirrors the local folder hierarchy. Notes with folderId=null go at the
-// repo root. The .md filename is derived from the note title.
-
-function buildFolderPath(folderId: string | null, folders: Folder[]): string {
-  if (!folderId) return ''
-  const byId = new Map(folders.map(f => [f.id, f]))
-  const segs: string[] = []
-  let cur: Folder | undefined = byId.get(folderId)
-  // Walk up to root; guard against cycles with a depth cap.
-  for (let i = 0; cur && i < 32; i++) {
-    if (cur.isDeleted) break
-    segs.unshift(sanitizeFilename(cur.name))
-    cur = cur.parentId ? byId.get(cur.parentId) : undefined
-  }
-  return segs.join('/')
-}
-
-// Repo-paths (e.g. `.obsidian/themes`) for every non-deleted local folder.
-// Used by the pull's directory-walking pass to skip dirs we already
-// materialised — without it we'd emit duplicate folderCreated entries on
-// every sync.
-function collectLocalFolderRepoPaths(folders: Folder[]): Set<string> {
-  const out = new Set<string>()
-  for (const f of folders) {
-    if (f.isDeleted) continue
-    const p = buildFolderPath(f.id, folders)
-    if (p) out.add(p)
-  }
-  return out
-}
-
-export function notePath(note: Note, folders: Folder[]): string {
-  const dir = buildFolderPath(note.folderId, folders)
-  const file = `${sanitizeFilename(note.title || 'Untitled')}.md`
-  return dir ? `${dir}/${file}` : file
-}
-
-// preserve-gitpath-on-push (the sanitizer-churn fix): the destination path a
-// PUSH writes a note to.
+// Thin barrel re-export for the post-split GitHub sync stack. Issue #77 split
+// the original ~1600-line `githubSync.ts` monolith into focused submodules
+// under `./githubSync/`:
 //
-// THE CHURN BUG: on clone a note's gitPath is set to the REAL remote path
-// (e.g. "R&D Work.md"). On push we used to ALWAYS re-derive the path from the
-// title via notePath(), which runs the title back through sanitizeFilename.
-// The old aggressive sanitizer stripped git-legal characters (`&`, `'`, `:`,
-// …), so the recomputed path ("RD Work.md") no longer matched the stored
-// gitPath ("R&D Work.md"). The push then deleted the real file + created a
-// stripped-name copy — rename churn on EVERY sync, permanently mangling the
-// user's filenames. (The relaxed sanitizer in sanitizeFilename.ts is the other
-// half of the fix; together they stop the drift.)
+//   - `syncClassify.ts` — the `PullClassification` discriminated-union the
+//     pull emits and the apply layer consumes.
+//   - `syncPull.ts`     — `pullFromGitHub`, `pullFromZipball`, and the
+//     zipball attachment side-channel cache.
+//   - `syncPush.ts`     — `syncToGitHub` and the upload-cache plumbing.
+//   - `internal.ts`     — path/serialization/encryption helpers shared by
+//     both halves.
 //
-// THE RULE:
-//   * NEW note (gitPath null/empty — never synced): derive a fresh path from
-//     the title via notePath() (relaxed sanitizer). It has no remote path yet.
-//   * SYNCED note whose title + folder STILL derive to its stored gitPath:
-//     use the stored gitPath VERBATIM. This is the no-churn path — even if any
-//     residual sanitizer difference existed, the cloned note stays pinned to
-//     the exact path the remote already has, so a freshly-cloned vault pushes
-//     nothing.
-//   * SYNCED note whose derived path DIFFERS from gitPath: the user genuinely
-//     RENAMED the title or MOVED the folder (updateNote/moveNoteToFolder change
-//     title/folderId but NOT gitPath — propagation is the push's job). We
-//     return the freshly-derived path so the move reaches the remote; the
-//     deletion loop (step 4) sha:null's the old gitPath. NOT breaking this is
-//     why we re-derive instead of blindly trusting gitPath.
-//
-// Folder names feed the derived path through buildFolderPath (also relaxed), so
-// the same principle covers folder-name churn: a synced note under a folder
-// like "Users & groups" derives back to its stored gitPath → no churn; a real
-// folder rename/move derives a new path → propagates.
-export function pushPath(note: Note, folders: Folder[]): string {
-  if (!note.gitPath) return notePath(note, folders)
-  // Synced note: trust the stored path unless a genuine rename/move means the
-  // derived path no longer matches it.
-  const derived = notePath(note, folders)
-  return derived === note.gitPath ? note.gitPath : derived
-}
-
-// Map common image extensions to MIME types so attachment pulls hand the
-// apply layer a properly-typed Blob. Unknown extensions fall back to
-// `application/octet-stream` — the file still round-trips, just without a
-// recognised type for browser previews.
-const MIME_BY_EXT: Record<string, string> = {
-  png: 'image/png',
-  jpg: 'image/jpeg',
-  jpeg: 'image/jpeg',
-  gif: 'image/gif',
-  webp: 'image/webp',
-  svg: 'image/svg+xml',
-  bmp: 'image/bmp',
-  avif: 'image/avif',
-}
-
-export function guessMimeFromPath(path: string): string {
-  const dotIdx = path.lastIndexOf('.')
-  if (dotIdx === -1) return 'application/octet-stream'
-  const ext = path.slice(dotIdx + 1).toLowerCase()
-  return MIME_BY_EXT[ext] ?? 'application/octet-stream'
-}
-
-// ── Note serialization ──────────────────────────────────────────────────────
-// We write the body verbatim — no YAML frontmatter. Tags now live as `#word`
-// patterns inline in the body, so there's nothing to round-trip in a header.
-// Round-trip identity uses the file path (Phase 4 pull matches by gitPath).
-//
-// IMPORTANT: we normalise to LF line endings + a single trailing newline so
-// our blob SHA matches what Obsidian (and most editors that follow the POSIX
-// "text files end in \n" convention) write for the same logical content. Without
-// this, every Obsidian-side save would re-touch the file and noteser would see
-// the trailing-newline difference as drift, re-uploading every blob on each
-// sync (the storm bug). See `normalizeForPush` for the canonical form.
-export function serializeNote(note: Note): string {
-  return normalizeForPush(note.content ?? '')
-}
-
-// Canonical wire form: CRLF → LF, ensure exactly one trailing \n, drop a
-// completely empty file's "newline" (an empty file is just empty bytes —
-// adding "\n" would create drift the other way).
-export function normalizeForPush(content: string): string {
-  if (content === '') return ''
-  const lf = content.replace(/\r\n/g, '\n')
-  return lf.endsWith('\n') ? lf : `${lf}\n`
-}
-
-// content-normalization-churn: "did the user actually edit this note since the
-// last sync?" — decided by a BYTE-EXACT comparison of the two bodies AFTER
-// canonicalisation, NOT by diffing blob SHAs.
-//
-// The blob-SHA path catches NORMALIZATION DRIFT: a note's stored body and the
-// exact remote blob it was pulled from can differ purely by line endings
-// (CRLF↔LF) or a trailing newline — never by anything the user typed. Hashing
-// the two raw forms yields different SHAs, so a SHA-only "changed?" test reports
-// a phantom edit and re-pushes a note the user never touched (the churn). Real
-// repro: a smart-punctuation note ("Don’t…" — U+2019/U+2014/U+2009/U+00A0) with
-// NO trailing newline whose baseline is the RAW (non-canonical) remote SHA —
-// e.g. a legacy note synced before gitLastPushedSha was pinned to the canonical
-// form, or a conflict-resolved note pinned to the raw remote SHA.
-//
-// Canonicalising BOTH sides (normalizeForPush) collapses exactly that drift —
-// CRLF→LF and trailing-newline — while preserving every byte the user can type,
-// INCLUDING all smart punctuation (curly quotes, em dashes, thin/non-breaking
-// spaces survive verbatim). So this returns true iff the only difference between
-// the two bodies is normalization → the note is UNEDITED.
-export function isUnchangedModuloNormalization(localContent: string, remoteContent: string): boolean {
-  return normalizeForPush(localContent) === normalizeForPush(remoteContent)
-}
-
-// ── Parser (Phase 4 pull) ───────────────────────────────────────────────────
-// We only support the YAML subset we ourselves emit / commonly see in
-// Obsidian vaults: `tags: [a, "b", c]` or `aliases: [Short, "Even Shorter"]`
-// on a single line. Anything else in the frontmatter is preserved into the
-// body so we don't silently destroy custom user metadata.
-export interface ParsedNote {
-  tags: string[]
-  aliases: string[]
-  body: string
-}
-
-// Parse a single-line YAML inline-array field (e.g. `tags: [a, "b", c]`) out
-// of the given frontmatter block. Returns [] when the field is absent or the
-// list is empty. Splits on commas, but ignores commas inside double quotes —
-// good enough for the formats we produce or encounter in real Obsidian vaults.
-function parseInlineArrayField(fmBlock: string, fieldName: string): string[] {
-  const re = new RegExp(`^${fieldName}:\\s*\\[(.*)\\]\\s*$`, 'm')
-  const lineMatch = fmBlock.match(re)
-  if (!lineMatch) return []
-  const inner = lineMatch[1].trim()
-  if (!inner) return []
-  const out: string[] = []
-  let cur = ''
-  let inQuote = false
-  for (let i = 0; i < inner.length; i++) {
-    const c = inner[i]
-    if (c === '"') { inQuote = !inQuote; continue }
-    if (c === ',' && !inQuote) {
-      const t = cur.trim()
-      if (t) out.push(t)
-      cur = ''
-      continue
-    }
-    cur += c
-  }
-  const t = cur.trim()
-  if (t) out.push(t)
-  return out
-}
-
-export function parseNote(raw: string): ParsedNote {
-  // No frontmatter — everything is body.
-  if (!raw.startsWith('---\n') && !raw.startsWith('---\r\n')) {
-    return { tags: [], aliases: [], body: raw }
-  }
-  // Find the closing delimiter starting at line 1.
-  const endMatch = raw.match(/\n---\r?\n/)
-  if (!endMatch || endMatch.index === undefined) return { tags: [], aliases: [], body: raw }
-
-  const fmBlock = raw.slice(4, endMatch.index)
-  const bodyStart = endMatch.index + endMatch[0].length
-  const body = raw.slice(bodyStart)
-
-  const tags = parseInlineArrayField(fmBlock, 'tags')
-  const aliases = parseInlineArrayField(fmBlock, 'aliases')
-  return { tags, aliases, body }
-}
-
-// ── Pull (Phase 4) ──────────────────────────────────────────────────────────
-
-// `adoptPath` (pull-dedupe-by-path): present only when this classification
-// is the result of the fallback reconciliation that matched a remote file to
-// an UNLINKED local note (gitPath null/empty, or a stale gitPath no longer in
-// the remote tree). The adopted note had no gitPath pointing at this file, so
-// apply must SET note.gitPath = adoptPath. For a normal gitPath-keyed match
-// the field is absent and apply leaves gitPath untouched.
-export type PullClassification =
-  // Local & remote agree, nothing to do. `adoptPath` is set when the match was
-  // a reconciled unlinked note that still needs its gitPath linked on apply.
-  | { kind: 'unchanged'; noteId: string; adoptPath?: string }
-  // Remote has a file with no matching local note yet — create one.
-  // progressive-clone: on a FIRST clone we emit these WITHOUT fetching the body
-  // (remoteContent '', tags [], body '') and set `shell: true`. The apply layer
-  // then creates a SHELL note (content '', contentLoaded false) so the sidebar
-  // populates instantly; the body streams in afterwards. `shell` is absent/false
-  // for an incremental pull's remoteCreated, which carries the real body as today.
-  | { kind: 'remoteCreated'; path: string; remoteSha: string; remoteContent: string; tags: string[]; body: string; shell?: boolean }
-  // Local exists, remote changed since our last push, local has NOT changed
-  // since last sync — accept the remote version.
-  | { kind: 'remoteUpdated'; noteId: string; remoteSha: string; remoteContent: string; tags: string[]; body: string; adoptPath?: string }
-  // We previously pushed this note, but the file is gone from the repo and
-  // we haven't edited it locally since — soft-delete it locally.
-  | { kind: 'remoteDeleted'; noteId: string }
-  // Both sides changed — let the user pick.
-  | {
-      kind: 'conflict'
-      noteId: string
-      path: string
-      localContent: string
-      remoteSha: string
-      remoteContent: string
-      remoteTags: string[]
-      remoteBody: string
-      adoptPath?: string
-    }
-  // Both sides changed but the line-level edits don't overlap, so we 3-way
-  // merged automatically. Apply writes the merged content + pins
-  // gitLastPushedSha to remoteSha so the next push uploads the union edit.
-  | {
-      kind: 'autoMerged'
-      noteId: string
-      remoteSha: string
-      mergedContent: string
-      adoptPath?: string
-    }
-  // Remote deleted the file but we edited it locally — degenerate conflict
-  // that's still asking the user a question, treat it as a conflict variant.
-  | {
-      kind: 'conflictDeleted'
-      noteId: string
-      path: string
-      localContent: string
-    }
-  // Binary attachment: remote has this file, local doesn't. Apply step fetches
-  // the bytes and writes them to IDB at the same path.
-  | { kind: 'attachmentCreated'; path: string; remoteSha: string; mime: string }
-  // Binary attachment: local + remote both have it but content differs. We
-  // treat remote as authoritative for v1 (no per-attachment three-way merge).
-  | { kind: 'attachmentUpdated'; path: string; remoteSha: string; mime: string }
-  // Directory the remote tree implies (via any file inside it) that we don't
-  // have locally. Materialise it as an empty Folder so the sidebar reflects
-  // the repo's structure — surfaces `.obsidian/` and similar dotfile dirs.
-  | { kind: 'folderCreated'; path: string }
-  // Vault settings file (vs8x) at `${settingsFolderPath}/settings.json`.
-  // The apply step compares remoteUpdatedAt to the local store's
-  // vaultSettingsUpdatedAt; if remote is newer it overwrites the local
-  // vault-tagged keys. Includes the raw hash so apply can update
-  // lastPushedHash atomically.
-  | {
-      kind: 'vaultSettingsUpdated'
-      path: string
-      remoteSha: string
-      remoteUpdatedAt: number
-      remoteVault: Record<string, unknown>
-      remoteHash: string
-    }
-  // vs8x-conflict: both sides have drifted since the last sync. Apply
-  // doesn't overwrite — instead it opens a modal where the user
-  // resolves key-by-key. The classification carries every differing
-  // key with its local + remote values so the modal can render
-  // without a second pull.
-  | {
-      kind: 'vaultSettingsConflict'
-      path: string
-      remoteSha: string
-      remoteUpdatedAt: number
-      remoteVault: Record<string, unknown>
-      remoteHash: string
-      localVault: Record<string, unknown>
-      // Just the differing keys (intersection of {present in either}
-      // where local !== remote). Saves the modal from re-diffing.
-      diffKeys: string[]
-    }
-
-export interface PullOutcome {
-  classifications: PullClassification[]
-  latestCommitSha: string
-}
-
-export async function pullFromGitHub(input: {
-  token: string
-  repo: SyncRepo
-  notes: Note[]
-  folders: Folder[]
-  // Repo paths the user explicitly deleted (folderStore.
-  // deletedFolderPaths). Step 1b skips dir-walk for any path
-  // matching one of these, OR nested inside one of these, so
-  // hidden/system folders the user removed don't auto-re-derive.
-  excludedFolderPaths?: string[]
-  // Vault settings file path (vs8x). When set, the pull looks for this
-  // path in the remote tree and emits a `vaultSettingsUpdated`
-  // classification if the remote's embedded updatedAt is newer than
-  // localVaultUpdatedAt. Pass null / undefined to skip settings pull.
-  vaultSettingsPath?: string | null
-  vaultSettingsLocalUpdatedAt?: number
-  // progressive-clone: when true the caller has determined this is a first
-  // clone (no local notes/folders to reconcile), so every remote .md is
-  // `remoteCreated`. We emit them as SHELLS (empty body, shell:true) WITHOUT
-  // fetching any blob bodies — the sidebar populates instantly and bodies
-  // stream in afterwards. Omit/false for incremental pulls so they classify +
-  // fetch normally (lazily, only what the diff needs).
-  isFirstClone?: boolean
-  // Legacy progress hook from the prefetch era — no longer fired by
-  // pullFromGitHub now that the first clone is shell-only (there is no blob
-  // prefetch to report). Kept on the interface for back-compat with callers
-  // that still pass it; harmless. The progressive background fill reports its
-  // own progress via the onPhase callback wired in useGitHubSync.
-  onBlobProgress?: (loaded: number, total: number) => void
-}): Promise<PullOutcome> {
-  const { token, repo, notes } = input
-  const excluded = input.excludedFolderPaths ?? []
-  const vaultSettingsPath = input.vaultSettingsPath ?? null
-  const vaultSettingsLocalUpdatedAt = input.vaultSettingsLocalUpdatedAt ?? 0
-  const isFirstClone = input.isFirstClone ?? false
-  const { owner, name, branch } = repo
-  const headSha = await getBranchRefSha(token, owner, name, branch)
-  const treeSha = await getCommitTreeSha(token, owner, name, headSha)
-  const remoteTree = await getTreeMap(token, owner, name, treeSha)
-
-  // Build the gitignore matcher BEFORE walking the tree so step 1's
-  // .md loop can short-circuit on ignored paths. The matcher is also
-  // reused for step 1b (folder derivation), 1c (attachments), and
-  // step 2 (orphan detection).
-  // Layered gitignore (gi9n):
-  //   1. baked-in OS-junk defaults (.DS_Store, Thumbs.db, *.tmp, *.swp)
-  //   2. the vault's remote `.gitignore` (if any)
-  //   3. the per-device overlay from settingsStore
-  // Each layer is appended after the previous so a later negation
-  // (e.g. `!keep.tmp` in the overlay) can un-ignore an earlier rule.
-  // We ALWAYS include the defaults — otherwise an empty remote with
-  // an empty overlay would lose the OS-junk fallback.
-  const { parseGitignore, DEFAULT_MATCHER, DEFAULT_IGNORE_LINES, GITIGNORE_PATH } = await import('./gitignore')
-  let gitignoreMatcher = DEFAULT_MATCHER
-  const gitignoreSha = remoteTree.get(GITIGNORE_PATH)
-  let remoteRaw = ''
-  if (gitignoreSha) {
-    try {
-      remoteRaw = await getBlobContent(token, owner, name, gitignoreSha)
-    } catch {
-      remoteRaw = ''
-    }
-  }
-  try {
-    const { useSettingsStore } = await import('@/stores/settingsStore')
-    const overlay = useSettingsStore.getState().localGitignoreOverlay || ''
-    const combined = [
-      DEFAULT_IGNORE_LINES.join('\n'),
-      remoteRaw,
-      overlay,
-    ].filter(s => s && s.trim().length > 0).join('\n')
-    gitignoreMatcher = parseGitignore(combined)
-  } catch {
-    // Defaults already applied above — pulls shouldn't die over a
-    // malformed ignore file or a missing store.
-  }
-
-  const out: PullClassification[] = []
-  const seenLocalIds = new Set<string>()
-
-  // progressive-clone: a FIRST clone NO LONGER prefetches blob bodies here.
-  // We walk the tree (already one call) for titles/paths/SHAs only and emit
-  // SHELL `remoteCreated` classifications with an EMPTY body. The sidebar then
-  // populates instantly from those shells; the bodies stream in afterwards via
-  // the background fill (useGitHubSync wires that up) and on-open. This keeps
-  // the prefetch map empty on a first clone — the classify loop's `shell`
-  // branch never calls loadRemote, so no per-blob network trips happen during
-  // the pull. Incremental pulls keep the lazy per-blob fetch unchanged.
-  //
-  // (The old behaviour pre-fetched ALL .md blobs with bounded concurrency. That
-  // still avoided the per-file watchdog blowout but blocked the sidebar on the
-  // whole download; the progressive shell approach is strictly faster to first
-  // paint.)
-  const prefetchedBlobs = new Map<string, string>()
-
-  // 1. Walk every remote .md file.
-  for (const [path, remoteSha] of remoteTree) {
-    if (!path.endsWith('.md')) continue
-    if (gitignoreMatcher.isIgnored(path)) continue
-    // Look up by gitPath in ALL notes (incl. soft-deleted). A
-    // soft-deleted note at the same path means the user explicitly
-    // wants this gone — we MUST NOT treat the remote file as a new
-    // creation and resurrect it. Push step 4 will emit the
-    // `sha: null` tree entry to actually delete it.
-    let localMatch = notes.find(n => n.gitPath === path)
-
-    if (localMatch && localMatch.isDeleted) {
-      // Pending deletion — skip the fetch + classification entirely.
-      // seenLocalIds includes it so the orphan-detection branch below
-      // doesn't double-count.
-      seenLocalIds.add(localMatch.id)
-      out.push({ kind: 'unchanged', noteId: localMatch.id })
-      continue
-    }
-
-    // progressive-clone CORE SAFETY GUARD: a SHELL note (body not yet loaded)
-    // must NEVER be classified as anything but `unchanged`. Its `content` is ''
-    // (a placeholder, not the real body), so computing a local blob SHA from it
-    // would produce the SHA of an empty file — which mismatches the real remote
-    // blob and would (a) re-classify the note as a local edit and (b) on push
-    // overwrite the real remote file with an empty body. We short-circuit BEFORE
-    // serializeNote / gitBlobSha / loadRemote so no body work happens at all.
-    // The background fill / on-open path will load the body and re-classify the
-    // note normally once `contentLoaded` flips true. (gitLastPushedSha is also
-    // pinned to remoteSha for shells, so even without this guard a SHELL whose
-    // remote is unchanged would read unchanged — but this guard makes it
-    // unconditional and avoids the empty-body SHA computation entirely.)
-    if (localMatch && localMatch.contentLoaded === false) {
-      seenLocalIds.add(localMatch.id)
-      out.push({ kind: 'unchanged', noteId: localMatch.id })
-      continue
-    }
-
-    // Fetch the remote content lazily — only when we need it. bke1:
-    // decrypt the envelope when encryption is on; throws VaultLockedError
-    // upstream if the user hasn't unlocked.
-    let remoteContent: string | null = null
-    const loadRemote = async () => {
-      if (remoteContent === null) {
-        // Prefer the blob we already pulled in the first-clone prefetch; only
-        // hit the network on a miss (incremental pulls, or a blob the prefetch
-        // didn't cover). decrypt runs here either way.
-        const raw = prefetchedBlobs.get(remoteSha) ?? await getBlobContent(token, owner, name, remoteSha)
-        remoteContent = await maybeDecryptFromPull(raw)
-      }
-      return remoteContent
-    }
-
-    // pull-dedupe-by-path: no gitPath-keyed match. Before declaring this a
-    // brand-new remote file (which apply would materialise as a NEW note), try
-    // a fallback reconciliation: is there an UNLINKED local note that logically
-    // IS this file? "Unlinked" = gitPath null/empty, OR a stale gitPath that is
-    // no longer present in this pull's remote tree. We match by the note's
-    // computed repo path (notePath) === this remote path. Adopting prevents the
-    // duplicate (the "two Temp notes" twin) for an unpushed local note or one
-    // whose gitPath was cleared (e.g. a conflictDeleted respawn).
-    //
-    // `adoptPath` is the path we'll thread onto the resulting classification so
-    // the apply layer links the adopted note's gitPath. Left undefined for a
-    // normal gitPath-keyed match.
-    let adoptPath: string | undefined
-    if (!localMatch) {
-      // rename-not-delete: an "unlinked" local note is one that has no gitPath
-      // pointing at a file in THIS remote tree — either never pushed, or its
-      // stored gitPath is stale (a rename/respawn left it behind). Such a note
-      // is a candidate to ADOPT this remote file.
-      const isUnlinked = (n: Note): boolean => {
-        const gp = n.gitPath
-        if (!gp) return true
-        return !remoteTree.has(gp)
-      }
-      // Path-form match: the note's computed repo path equals this remote path.
-      const pathCandidates = notes.filter(n => {
-        if (n.isDeleted) return false
-        if (seenLocalIds.has(n.id)) return false
-        if (notePath(n, input.folders) !== path) return false
-        return isUnlinked(n)
-      })
-
-      if (pathCandidates.length === 1) {
-        localMatch = pathCandidates[0]
-        adoptPath = path
-      } else if (pathCandidates.length > 1) {
-        // Ambiguous — be conservative. Only adopt if EXACTLY ONE candidate's
-        // serialized blob SHA equals the remote SHA (a clean, identical adopt
-        // we can make with confidence). Otherwise we refuse to guess and fall
-        // through to remoteCreated rather than risk merging the wrong note.
-        const shaMatches: Note[] = []
-        for (const c of pathCandidates) {
-          const sha = await gitBlobSha(serializeNote(c))
-          if (sha === remoteSha) shaMatches.push(c)
-        }
-        if (shaMatches.length === 1) {
-          localMatch = shaMatches[0]
-          adoptPath = path
-        }
-      }
-
-      // rename-not-delete CONTENT-HASH ADOPTION: the path FORM differs (e.g. the
-      // note's stored gitPath/folder names are dash-form, the remote reverted to
-      // space-form), so notePath() no longer equals the remote path and the
-      // path-match above found nothing. Fall back to matching by CONTENT: an
-      // unlinked local note whose serialized blob SHA equals this remote blob
-      // (or whose last-pushed SHA equals it) logically IS this file under a new
-      // name. Adopting it (gitPath := remotePath) prevents the catastrophic
-      // "old-path deleted + new-path created" misread that soft-deletes the note
-      // and then deletes the real remote file. We only adopt on an UNAMBIGUOUS
-      // single content match, mirroring the conservative SHA tiebreak above.
-      if (!localMatch) {
-        const hashMatches: Note[] = []
-        for (const n of notes) {
-          if (n.isDeleted) continue
-          if (seenLocalIds.has(n.id)) continue
-          if (!isUnlinked(n)) continue
-          // Content-hash adoption is for RENAMES: a note that was PUSHED before
-          // (so it has a baseline gitPath + last-pushed SHA) whose stored path
-          // FORM no longer matches the remote. A never-pushed note (no gitPath
-          // AND no gitLastPushedSha) is the job of the notePath path-form match
-          // above — adopting it here by raw content equality is both unnecessary
-          // and riskier (it could vacuum up an unrelated local draft that merely
-          // happens to share bytes). Require a push lineage to qualify.
-          if (!n.gitPath && !n.gitLastPushedSha) continue
-          // Cheap check first: the note's recorded last-pushed SHA already
-          // equals this remote blob (it was pushed as this exact content).
-          if (n.gitLastPushedSha === remoteSha) { hashMatches.push(n); continue }
-          // Otherwise hash the note's current serialized content.
-          const sha = await gitBlobSha(serializeNote(n))
-          if (sha === remoteSha) hashMatches.push(n)
-        }
-        if (hashMatches.length === 1) {
-          localMatch = hashMatches[0]
-          adoptPath = path
-        }
-      }
-    }
-
-    if (!localMatch) {
-      // progressive-clone: on a FIRST clone, emit a SHELL — no body fetch. The
-      // apply layer materialises a placeholder note (content '', contentLoaded
-      // false) so the sidebar populates instantly; the body streams in later.
-      if (isFirstClone) {
-        out.push({ kind: 'remoteCreated', path, remoteSha, remoteContent: '', tags: [], body: '', shell: true })
-        continue
-      }
-      const content = await loadRemote()
-      const parsed = parseNote(content)
-      out.push({ kind: 'remoteCreated', path, remoteSha, remoteContent: content, tags: parsed.tags, body: parsed.body })
-      continue
-    }
-
-    seenLocalIds.add(localMatch.id)
-    const localContent = serializeNote(localMatch)
-    const localBlobSha = await gitBlobSha(localContent)
-
-    if (localBlobSha === remoteSha) {
-      // Even for a clean `unchanged`, an adopted note still needs its gitPath
-      // linked on apply — thread adoptPath through.
-      out.push({ kind: 'unchanged', noteId: localMatch.id, ...(adoptPath ? { adoptPath } : {}) })
-      continue
-    }
-
-    // Two distinct SHAs (the two-SHA split that fixes silent conflict loss):
-    //   - localChanged compares the canonical LOCAL bytes against the local
-    //     baseline (gitLastPushedSha). For a frontmatter note this baseline is
-    //     the SHA of the TRANSFORMED body the app stores, so an untouched note
-    //     hashes back to it → localChanged = false (no phantom drift).
-    //   - remoteChanged compares the current remote blob SHA against the REMOTE
-    //     blob we last synced against (gitRemoteBaseSha). That blob is the raw
-    //     remote file (with frontmatter), which is exactly what the three-way
-    //     ancestor fetch needs.
-    // Un-migrated notes (synced before gitRemoteBaseSha existed) have no remote
-    // base — fall back to gitLastPushedSha to preserve the prior behaviour
-    // until their next sync rewrites both fields.
-    const localBaseline = localMatch.gitLastPushedSha ?? null
-    const remoteBase = localMatch.gitRemoteBaseSha ?? localMatch.gitLastPushedSha ?? null
-    const remoteChanged = remoteBase !== remoteSha
-    const localChanged = localBaseline !== localBlobSha
-
-    if (!remoteChanged && !localChanged) {
-      // Neither side moved since the last sync. The raw remote blob SHA still
-      // differs from our canonical local SHA (that's normal for a frontmatter
-      // note: we store a transformed body) — but nothing has actually changed,
-      // so this is `unchanged`. Without this branch a frontmatter note would
-      // fall through to "remoteUnchanged + localChanged → push", silently
-      // re-pushing on every sync (the storm) AND never settling to unchanged.
-      out.push({ kind: 'unchanged', noteId: localMatch.id, ...(adoptPath ? { adoptPath } : {}) })
-    } else if (remoteChanged && !localChanged) {
-      const content = await loadRemote()
-      const parsed = parseNote(content)
-      out.push({ kind: 'remoteUpdated', noteId: localMatch.id, remoteSha, remoteContent: content, tags: parsed.tags, body: parsed.body, ...(adoptPath ? { adoptPath } : {}) })
-    } else if (remoteChanged && localChanged) {
-      const content = await loadRemote()
-      const parsed = parseNote(content)
-
-      // Try a line-level 3-way merge before bothering the user. If the local
-      // and remote edits don't overlap line-wise we can auto-merge and the
-      // user never sees the conflict tab. The common ancestor is the REMOTE
-      // blob we last synced against (`gitRemoteBaseSha`, fetchable via
-      // getBlobContent) — NOT gitLastPushedSha, which is the SHA of the
-      // transformed local bytes and may not exist as a remote blob at all.
-      // Anything that goes wrong (no ancestor sha, blob GC'd, network hiccup,
-      // overlapping edits) falls back to the existing manual conflict flow.
-      let autoMerged: string | null = null
-      if (remoteBase) {
-        try {
-          const ancestorRaw = await getBlobContent(token, owner, name, remoteBase)
-          const ancestor = await maybeDecryptFromPull(ancestorRaw)
-          const merged = threeWayMerge(ancestor, localContent, content)
-          if (merged.ok) autoMerged = merged.merged
-        } catch {
-          // Swallow — fall through to conflict.
-        }
-      }
-
-      if (autoMerged !== null) {
-        out.push({
-          kind: 'autoMerged',
-          noteId: localMatch.id,
-          remoteSha,
-          mergedContent: autoMerged,
-          ...(adoptPath ? { adoptPath } : {}),
-        })
-      } else {
-        out.push({
-          kind: 'conflict',
-          noteId: localMatch.id,
-          path,
-          localContent,
-          remoteSha,
-          remoteContent: content,
-          remoteTags: parsed.tags,
-          remoteBody: parsed.body,
-          ...(adoptPath ? { adoptPath } : {}),
-        })
-      }
-    }
-    // remoteUnchanged + localChanged → handled by the push phase, nothing here.
-  }
-
-  // 1b. Empty / non-syncable directories the remote implies. We classify
-  // every parent directory of every blob; the apply step calls
-  // ensureFolderPath on each, so dotfile dirs like `.obsidian/` and
-  // `.obsidian/themes/` show in the sidebar even though we don't pull their
-  // file contents. The `attachments/` tree is excluded — it stays rendered
-  // by the sidebar's synthetic folder, not as a real Folder entity.
-  //
-  // Skip "dying" paths: blobs whose matched local note is soft-deleted, OR
-  // whose matched note's desired path differs from its stale gitPath (the
-  // note moved — push will sha:null the old path). Without this skip,
-  // deleting a folder re-derives it on the very next pull because the
-  // moved-to-root notes still carry the old gitPath. See bug "delete
-  // hidden folder, it appears again."
-  const localFolderPaths = collectLocalFolderRepoPaths(input.folders)
-  const pendingRemovedPaths = new Set<string>()
-  for (const [path] of remoteTree) {
-    if (!path.endsWith('.md')) continue
-    const localMatch = notes.find(n => n.gitPath === path)
-    if (!localMatch) continue
-    if (localMatch.isDeleted) {
-      pendingRemovedPaths.add(path)
-      continue
-    }
-    const desiredPath = notePath(localMatch, input.folders)
-    if (desiredPath !== path) pendingRemovedPaths.add(path)
-  }
-  // Helper: is `dir` either tombstoned itself or nested inside a
-  // tombstoned ancestor? Linear scan over `excluded` — list is tiny
-  // (one entry per explicit user-delete) so a Set+walk isn't worth it.
-  const isExcluded = (dir: string): boolean => {
-    for (const ex of excluded) {
-      if (dir === ex) return true
-      if (dir.startsWith(`${ex}/`)) return true
-    }
-    return false
-  }
-  const seenDirPaths = new Set<string>()
-  for (const [path] of remoteTree) {
-    if (pendingRemovedPaths.has(path)) continue
-    // Skip dir-walk for ignored files — they shouldn't surface their
-    // parent directories either. Otherwise an ignored `.DS_Store` in
-    // a subfolder would still cause the subfolder to be derived.
-    if (gitignoreMatcher.isIgnored(path)) continue
-    let cur = path
-    while (true) {
-      const lastSlash = cur.lastIndexOf('/')
-      if (lastSlash === -1) break
-      cur = cur.slice(0, lastSlash)
-      if (!cur) break
-      if (seenDirPaths.has(cur)) break
-      seenDirPaths.add(cur)
-      if (localFolderPaths.has(cur)) continue
-      if (isExcluded(cur)) continue
-      if (gitignoreMatcher.isIgnored(cur, true)) continue
-      out.push({ kind: 'folderCreated', path: cur })
-    }
-  }
-
-  // 1d. Vault settings file (vs8x). If the caller is opting in
-  // (vaultSettingsPath set) and the remote has the file, fetch it,
-  // parse, and emit a classification iff its embedded updatedAt is
-  // strictly newer than ours. Equal or older → skip (LWW with us as
-  // the tiebreaker so a clean re-push doesn't clobber a slightly
-  // newer local edit).
-  if (vaultSettingsPath) {
-    const remoteSettingsSha = remoteTree.get(vaultSettingsPath)
-    if (remoteSettingsSha) {
-      try {
-        const raw = await getBlobContent(token, owner, name, remoteSettingsSha)
-        const { parseVaultSettings, vaultSettingsHash, pickVaultSlice, serializeVaultSettings } = await import('./vaultSettings')
-        const parsed = parseVaultSettings(raw)
-        if (parsed && parsed.updatedAt > vaultSettingsLocalUpdatedAt) {
-          const remoteHash = vaultSettingsHash(raw)
-          // vs8x-conflict: are local AND remote both dirty since the
-          // last sync? If localHash !== lastPushedHash the user has
-          // edits we never pushed — overlaying the remote silently
-          // would clobber them. Open a modal instead.
-          const { useSettingsStore } = await import('@/stores/settingsStore')
-          const settingsState = useSettingsStore.getState()
-          const localVaultSlice = pickVaultSlice(settingsState)
-          const localCanonical = serializeVaultSettings(localVaultSlice, settingsState.vaultSettingsUpdatedAt || 0)
-          const localHash = vaultSettingsHash(localCanonical)
-          const localDirty = localHash !== (settingsState.vaultSettingsLastPushedHash || '')
-
-          const remoteVaultObj = parsed.vault as Record<string, unknown>
-          const localVaultObj = localVaultSlice as Record<string, unknown>
-
-          if (localDirty) {
-            // Build the diff key list — every key whose local + remote
-            // values differ. Both sides whitelist to VAULT_SETTING_KEYS
-            // already, so the comparison stays small.
-            const keys = new Set<string>([
-              ...Object.keys(localVaultObj),
-              ...Object.keys(remoteVaultObj),
-            ])
-            const diffKeys: string[] = []
-            for (const k of keys) {
-              if (JSON.stringify(localVaultObj[k]) !== JSON.stringify(remoteVaultObj[k])) {
-                diffKeys.push(k)
-              }
-            }
-            if (diffKeys.length > 0) {
-              out.push({
-                kind: 'vaultSettingsConflict',
-                path: vaultSettingsPath,
-                remoteSha: remoteSettingsSha,
-                remoteUpdatedAt: parsed.updatedAt,
-                remoteVault: remoteVaultObj,
-                remoteHash,
-                localVault: localVaultObj,
-                diffKeys,
-              })
-            }
-            // If diffKeys is empty the values match anyway → fall
-            // through to vaultSettingsUpdated which is a cheap no-op.
-            else {
-              out.push({
-                kind: 'vaultSettingsUpdated',
-                path: vaultSettingsPath,
-                remoteSha: remoteSettingsSha,
-                remoteUpdatedAt: parsed.updatedAt,
-                remoteVault: remoteVaultObj,
-                remoteHash,
-              })
-            }
-          } else {
-            // Clean local → remote wins as before.
-            out.push({
-              kind: 'vaultSettingsUpdated',
-              path: vaultSettingsPath,
-              remoteSha: remoteSettingsSha,
-              remoteUpdatedAt: parsed.updatedAt,
-              remoteVault: remoteVaultObj,
-              remoteHash,
-            })
-          }
-        }
-      } catch {
-        // Bad JSON / network blip — skip rather than fail the entire
-        // pull. The settings file is a soft signal; notes still sync.
-      }
-    }
-  }
-
-  // 1c. Binary attachments under `attachments/`. Compare each remote entry
-  // against the local IDB store; queue creates/updates so syncApply can fetch
-  // the bytes lazily (each blob fetch is its own API call, so we only pay
-  // for ones the user actually needs).
-  const localAttachmentPaths = new Set(await listAttachmentPaths())
-  for (const [path, remoteSha] of remoteTree) {
-    if (!isAttachmentPath(path)) continue
-    if (gitignoreMatcher.isIgnored(path)) continue
-    // Best-effort MIME guess from extension — the apply step uses this to
-    // build the Blob. Falls back to octet-stream.
-    const mime = guessMimeFromPath(path)
-    if (!localAttachmentPaths.has(path)) {
-      out.push({ kind: 'attachmentCreated', path, remoteSha, mime })
-      continue
-    }
-    const localSha = await getAttachmentGitSha(path)
-    if (localSha && localSha !== remoteSha) {
-      out.push({ kind: 'attachmentUpdated', path, remoteSha, mime })
-    }
-  }
-
-  // 2. Local notes that had a gitPath but are missing from the remote tree.
-  for (const note of notes) {
-    if (note.isDeleted || !note.gitPath || seenLocalIds.has(note.id)) continue
-    if (remoteTree.has(note.gitPath)) continue
-    // Was it deleted on the remote?
-    const lastPushed = note.gitLastPushedSha ?? null
-    const localContent = serializeNote(note)
-    const localBlobSha = await gitBlobSha(localContent)
-    if (lastPushed && lastPushed === localBlobSha) {
-      // We haven't touched it locally since the last push → accept the delete.
-      out.push({ kind: 'remoteDeleted', noteId: note.id })
-    } else if (lastPushed) {
-      // Remote deleted, local has edits since the last sync → conflict.
-      out.push({ kind: 'conflictDeleted', noteId: note.id, path: note.gitPath, localContent })
-    }
-    // No lastPushed → this note was never actually synced (clear stale gitPath
-    // and let push re-create it). Treat as remoteDeleted to just clear state.
-    else {
-      out.push({ kind: 'remoteDeleted', noteId: note.id })
-    }
-  }
-
-  return { classifications: out, latestCommitSha: headSha }
-}
-
-// ── Bulk first-clone (zipball fast path) ────────────────────────────────────
-//
-// `pullFromGitHub` is correct but does one blob fetch per file that differs
-// from local. On a first connection to a large vault (thousands of files) we
-// already know everything is `remoteCreated`, so the per-file API trip is
-// pure waste — we'd burn rate limit and minutes of wall time. The zipball
-// endpoint hands us the whole repo as a single zip download, which the
-// browser already follows past a redirect on its own and which doesn't get
-// charged against the primary API rate limit the way blob reads do.
-//
-// We still need authoritative blob SHAs to seed `gitLastPushedSha`. Computing
-// them locally via `gitBlobSha` produces the same hash git would (it's the
-// same SHA-1 of `blob <len>\0<content>`), so a separate tree fetch isn't
-// necessary.
-export async function pullFromZipball(input: {
-  token: string
-  repo: SyncRepo
-  // Phase hint so the caller can surface "Downloading vault (retrying)…" when
-  // a corrupted/truncated archive triggers a re-download. Optional — the retry
-  // loop works regardless.
-  onPhase?: (msg: string) => void
-}): Promise<PullOutcome> {
-  const { token, repo, onPhase } = input
-  const { owner, name, branch } = repo
-
-  // The ref is cheap and we need it for `latestCommitSha` regardless — fetch
-  // it once up front, independent of the archive retry loop below.
-  const headSha = await getBranchRefSha(token, owner, name, branch)
-
-  // Lazy-load jszip — only callers of pullFromZipball pay the
-  // ~140kB cost. The rest of the sync flow (push, regular pull via
-  // Git Data API) never touches it.
-  const { default: JSZip } = await import('jszip')
-
-  // Download + parse the archive with a short retry loop. On a large vault
-  // over a flaky mobile connection the single big zip download sometimes
-  // arrives truncated, so JSZip throws "Corrupted zip: can't find end of
-  // central directory" (or fetchZipball's own Content-Length guard fires).
-  // Both are transient — re-downloading usually succeeds within a couple of
-  // tries, so we do that automatically instead of making the user tap Retry.
-  // Only after exhausting the attempts do we surface the error (→ existing
-  // toast with Retry). The per-request 180s timeout inside fetchZipball is
-  // unchanged; this loop wraps whole-download attempts, not single requests.
-  const MAX_ATTEMPTS = 3
-  const BACKOFF_MS = [500, 1_000]
-  let zip: JSZip
-  let attempt = 0
-  for (;;) {
-    try {
-      const zipBuffer = await fetchZipball(token, owner, name, branch)
-      zip = await JSZip.loadAsync(zipBuffer)
-      break
-    } catch (err) {
-      attempt++
-      if (attempt >= MAX_ATTEMPTS) throw err
-      // Short backoff before re-downloading, and tell the UI we're retrying —
-      // with the attempt number, so a flaky large-vault download reads as
-      // progress, not a stall. (attempt was just incremented; the next try is
-      // attempt + 1.)
-      onPhase?.(`Vault download incomplete, retrying (${attempt + 1} of ${MAX_ATTEMPTS})…`)
-      const delay = BACKOFF_MS[attempt - 1] ?? BACKOFF_MS[BACKOFF_MS.length - 1]
-      await new Promise<void>((resolve) => setTimeout(resolve, delay))
-    }
-  }
-  const classifications: PullClassification[] = []
-
-  // The zipball wraps every entry in a top-level directory named
-  // `<owner>-<repo>-<short-sha>/`, so we strip the first path segment.
-  const entries: Array<{ rel: string; file: JSZip.JSZipObject }> = []
-  zip.forEach((rel, file) => {
-    if (file.dir) return
-    // Pull both .md notes and binary files under the attachments folder
-    // (configured or historical default). Anything else in the repo (root
-    // README, .github/, etc.) is ignored on the first clone.
-    if (rel.endsWith('.md')) {
-      entries.push({ rel, file })
-      return
-    }
-    const slashIdx = rel.indexOf('/')
-    if (slashIdx !== -1 && isAttachmentPath(rel.slice(slashIdx + 1))) {
-      entries.push({ rel, file })
-    }
-  })
-
-  for (const { rel, file } of entries) {
-    const slashIdx = rel.indexOf('/')
-    if (slashIdx === -1) continue
-    const path = rel.slice(slashIdx + 1)
-
-    if (path.endsWith('.md')) {
-      const raw = await file.async('string')
-      // bke1: zipball blobs were written in the encrypted wire form. We
-      // compute the remoteSha against that wire form (matches what
-      // GitHub stored) but feed parseNote the decrypted body.
-      const remoteSha = await gitBlobSha(raw)
-      const content = await maybeDecryptFromPull(raw)
-      const parsed = parseNote(content)
-
-      classifications.push({
-        kind: 'remoteCreated',
-        path,
-        remoteSha,
-        remoteContent: content,
-        tags: parsed.tags,
-        body: parsed.body,
-      })
-      continue
-    }
-
-    if (isAttachmentPath(path)) {
-      const bytes = await file.async('uint8array')
-      const remoteSha = await gitBlobShaBytes(bytes)
-      const mime = guessMimeFromPath(path)
-      classifications.push({ kind: 'attachmentCreated', path, remoteSha, mime })
-      // pullFromZipball already has the bytes in memory — stash them so the
-      // apply step doesn't issue a redundant per-blob fetch against the API.
-      attachmentBytesByPath.set(path, { bytes, mime })
-      continue
-    }
-  }
-
-  return { classifications, latestCommitSha: headSha }
-}
-
-// Side-channel cache: pullFromZipball already has the bytes in memory after
-// reading the zip, so we stash them here and the apply layer (or anyone
-// calling getZipballAttachmentBytes) can grab them without re-downloading.
-// Cleared after applyAttachmentClassifications consumes them.
-const attachmentBytesByPath = new Map<string, { bytes: Uint8Array; mime: string }>()
-
-export function takeZipballAttachmentBytes(
-  path: string,
-): { bytes: Uint8Array; mime: string } | null {
-  const entry = attachmentBytesByPath.get(path)
-  if (!entry) return null
-  attachmentBytesByPath.delete(path)
-  return entry
-}
-
-// ── Sync orchestrator ───────────────────────────────────────────────────────
-
-export type PushProgress =
-  | { phase: 'computing' }
-  | { phase: 'uploading-blobs'; uploaded: number; total: number; skipped: number }
-  | { phase: 'creating-tree' }
-  | { phase: 'creating-commit' }
-  | { phase: 'updating-ref' }
-  | { phase: 'done' }
-
-export interface SyncInput {
-  token: string
-  repo: SyncRepo
-  notes: Note[]
-  folders: Folder[]
-  // Optional override for the GitHub commit message. When omitted we
-  // auto-generate "Sync from Noteser (N changes)". Used by the
-  // obsidian-git-style commit-message box (vscg).
-  commitMessage?: string
-  // Vault settings bundle (vs8x). Caller serializes the vault slice +
-  // hash + last-pushed-hash; we include the file in the push tree when
-  // either the hash changed locally or the file is missing remotely.
-  // Pass undefined to skip settings sync entirely (e.g. user cleared
-  // settingsFolderPath).
-  vaultSettings?: {
-    path: string
-    content: string
-    contentHash: string
-    lastPushedHash: string
-  }
-  // Pending vault `.gitignore` edit from the in-app editor (gi9n
-  // Settings UI). When provided, the push tree includes a `.gitignore`
-  // blob with this content IF it differs from the current remote file.
-  // Pass undefined when the user hasn't touched the editor — the
-  // remote file is left alone.
-  vaultGitignoreDraft?: string | null
-  // Push-progress hook so the UI can surface "uploading 47 / 200 blobs"
-  // and tell the user which step failed when an error bubbles. Optional.
-  onProgress?: (event: PushProgress) => void
-}
-
-// In-memory cache of blob SHAs we've already uploaded to GitHub in this
-// tab session. Git blob SHAs are content-addressable, so a hit here
-// means GitHub already has that content — skip the redundant network
-// round-trip. Survives across syncToGitHub calls within the tab but is
-// cleared when the user reloads. Indexed per-repo so two different
-// vaults don't share state.
-const uploadedBlobShaCache = new Map<string, Set<string>>()
-
-function repoCacheKey(repo: SyncRepo): string {
-  return `${repo.owner}/${repo.name}#${repo.branch}`
-}
-
-function getUploadedShas(repo: SyncRepo): Set<string> {
-  const key = repoCacheKey(repo)
-  let set = uploadedBlobShaCache.get(key)
-  if (!set) {
-    set = new Set()
-    uploadedBlobShaCache.set(key, set)
-  }
-  return set
-}
-
-/** Test hook. Drops the in-memory upload cache. */
-export function _resetUploadedShaCache(): void {
-  uploadedBlobShaCache.clear()
-}
-
-export type GitPathUpdate = {
-  noteId: string
-  gitPath: string | null
-  // Blob SHA of what we just pushed for this note. Null when the note's file
-  // was deleted in this commit. After a push the remote file IS
-  // serializeNote(note), so this is BOTH the canonical local baseline and the
-  // remote merge base — gitRemoteBaseSha below is set to the same value.
-  gitLastPushedSha: string | null
-  // The remote merge-base SHA to persist alongside gitLastPushedSha. Coincides
-  // with gitLastPushedSha after a push (the pushed blob is the new remote
-  // file). Null when the file was deleted.
-  gitRemoteBaseSha: string | null
-}
-
-export interface SyncOutcome {
-  result: SyncResult
-  // The caller applies these updates to the note store so each note remembers
-  // where it lives in the repo (and which ones are no longer there).
-  pathUpdates: GitPathUpdate[]
-  // Set when the vault settings file was included in the push (or was
-  // already up to date with `lastPushedHash`). Caller should persist
-  // this so subsequent pushes skip when content-equal.
-  vaultSettingsHashPushed?: string
-  // True when this push wrote the vault `.gitignore` (i.e. the caller's
-  // `vaultGitignoreDraft` differed from remote and we uploaded it).
-  // Caller should clear the draft + refresh its snapshot to the pushed
-  // content so the next sync doesn't try to push again.
-  vaultGitignorePushed?: boolean
-}
-
-export async function syncToGitHub(input: SyncInput): Promise<SyncOutcome> {
-  const { token, repo, notes, folders, commitMessage, vaultSettings, vaultGitignoreDraft, onProgress } = input
-  const { owner, name, branch } = repo
-  const uploadedShas = getUploadedShas(repo)
-  onProgress?.({ phase: 'computing' })
-
-  // 1. Compute desired files for every active note.
-  // progressive-clone: EXCLUDE shells (contentLoaded === false) outright. A
-  // shell's `content` is '' (a placeholder, not the real body), so it must
-  // never participate in the push: it would (a) upload an empty blob over the
-  // real remote file, and (b) — if we only skipped it from `desired` — get
-  // detected as a missing path by the deletion loop (step 4) and emit a
-  // `sha: null` DELETE for the real file. Dropping it from activeNotes here
-  // keeps it out of BOTH paths; the shell stays exactly as the remote has it
-  // until its body loads and contentLoaded flips true. This is the push-side
-  // half of the #1 sync-safety rule (the pull-side half is the classifier
-  // guard in pullFromGitHub).
-  const activeNotes = notes.filter(n => !n.isDeleted && n.contentLoaded !== false)
-  const desired = new Map<string, { content: string; note: Note }>()
-  for (const note of activeNotes) {
-    // preserve-gitpath-on-push: a synced note pushes to its EXISTING gitPath
-    // verbatim; only a new note (no gitPath) derives a fresh path from its
-    // title. This is what stops the sanitizer-relax churn — a cloned note keeps
-    // the exact remote path even when the title contains characters the
-    // sanitizer would alter. See pushPath() for the rename rationale.
-    const path = pushPath(note, folders)
-    const content = serializeNote(note)
-    // If two notes resolve to the same path (e.g. duplicate titles in same
-    // folder), the later one in the array wins. Acceptable for v1.
-    desired.set(path, { content, note })
-  }
-
-  // 2. Fetch the current branch state.
-  const parentCommitSha = await getBranchRefSha(token, owner, name, branch)
-  const baseTreeSha = await getCommitTreeSha(token, owner, name, parentCommitSha)
-  const remoteTree = await getTreeMap(token, owner, name, baseTreeSha)
-
-  // Layered gitignore matcher for push (gi9n): defaults + remote +
-  // local overlay. Same composition as the pull side so push and pull
-  // agree on what to skip.
-  const { parseGitignore, DEFAULT_MATCHER, DEFAULT_IGNORE_LINES, GITIGNORE_PATH } = await import('./gitignore')
-  let pushMatcher = DEFAULT_MATCHER
-  const remoteGitignoreSha = remoteTree.get(GITIGNORE_PATH)
-  let pushRemoteRaw = ''
-  if (remoteGitignoreSha) {
-    try {
-      pushRemoteRaw = await getBlobContent(token, owner, name, remoteGitignoreSha)
-    } catch {
-      pushRemoteRaw = ''
-    }
-  }
-  try {
-    const { useSettingsStore } = await import('@/stores/settingsStore')
-    const overlay = useSettingsStore.getState().localGitignoreOverlay || ''
-    const combined = [
-      DEFAULT_IGNORE_LINES.join('\n'),
-      pushRemoteRaw,
-      overlay,
-    ].filter(s => s && s.trim().length > 0).join('\n')
-    pushMatcher = parseGitignore(combined)
-  } catch {
-    // Defaults already applied above.
-  }
-
-  // 3. Build tree entries for changes only.
-  const entries: GitTreeEntry[] = []
-  const pathUpdates: GitPathUpdate[] = []
-  let created = 0
-  let updated = 0
-  let deleted = 0
-
-  // We snapshot per-note pushed content so the editor's gutter diff
-  // (109) can compare against it. The IDB write is fire-and-forget
-  // outside the loop — collect first, write after the loop ends to
-  // avoid serial awaits per note.
-  const lastPushedToSnapshot: Array<{ noteId: string; content: string }> = []
-
-  // Pre-pass: classify every desired path into "skip" (remote already
-  // has this SHA, or our in-tab cache says we uploaded it) vs "needs
-  // upload". Keeping the pre-pass separate lets us emit a stable
-  // `total` to the progress callback.
-  //
-  // push-only-real-edits: the upload decision must reflect a GENUINE local
-  // edit, not just a wire-form mismatch against the remote blob. The remote
-  // blob can be in a NON-CANONICAL shape (e.g. an imported Obsidian vault with
-  // no trailing newline, or frontmatter we strip on store), so `remoteSha`
-  // routinely differs from our canonical `localSha` even when the user never
-  // touched the note. Uploading on that mismatch rewrites the user's vault into
-  // noteser's canonical form on every sync (the churn bug).
-  //
-  // The authoritative "did the user edit this?" signal is the PLAINTEXT
-  // canonical SHA (`plainSha = gitBlobSha(serializeNote(note))`) vs the note's
-  // `gitLastPushedSha` baseline (set by syncApply/backgroundFill to the
-  // canonical SHA as of the last sync). When they match the body is byte-equal
-  // to what we last synced → NO genuine edit. A null baseline means a
-  // new/never-synced note → MUST push. We use `plainSha` (NOT the wire/encrypted
-  // sha) ONLY for this change decision; `localSha` (wire) is still what we
-  // hash, upload and dedupe against `uploadedShas`/`remoteSha`.
-  interface NoteBlobPlan { path: string; content: string; note: Note; localSha: string; remoteSha: string | undefined; locallyChanged: boolean }
-  const noteBlobPlan: NoteBlobPlan[] = []
-  for (const [path, { content, note }] of desired) {
-    if (pushMatcher.isIgnored(path)) continue
-    // bke1: when encryption is enabled, the wire form is what we hash
-    // and upload. plaintext stays available for the editor's gutter
-    // snapshot (it compares against the unencrypted body).
-    const wireContent = await maybeEncryptForPush(content)
-    const localSha = await gitBlobSha(wireContent)
-    // Plaintext canonical SHA for the genuine-edit decision. `content` here is
-    // the canonical plaintext (serializeNote output) BEFORE encryption.
-    const plainSha = await gitBlobSha(content)
-    const baseline = note.gitLastPushedSha ?? null
-    const remoteSha = remoteTree.get(path)
-    // Primary signal: the canonical SHA equals the baseline → byte-identical to
-    // what we last synced → not a real edit. A null baseline is a brand-new note.
-    let locallyChanged = baseline === null || plainSha !== baseline
-
-    // content-normalization-churn: the canonical-SHA test above mis-fires for a
-    // note whose baseline is the RAW (non-canonical) remote blob rather than the
-    // canonical serialisation — a LEGACY note (synced before gitLastPushedSha was
-    // pinned to the canonical form) or a conflict-resolved note. There the remote
-    // blob lacks a trailing newline (or uses CRLF), so plainSha (canonical, with
-    // \n) != baseline (raw) even though the user never typed a thing — the churn.
-    //
-    // Disambiguate with a BYTE-EXACT comparison: when the baseline IS the current
-    // remote blob (baseline === remoteSha), the note was last synced AS that exact
-    // blob and its gitPath hasn't moved. Fetch the remote body ONCE and compare it
-    // to the local body modulo normalization. If they match, the difference is
-    // pure CRLF/trailing-newline drift → UNEDITED → suppress (leave the user's
-    // original non-canonical bytes on the remote untouched, exactly like the
-    // canonical-baseline path does). This network read only happens for the legacy
-    // drift case (SHA says changed AND baseline names the live remote blob), so the
-    // hot path (canonical baseline, or a genuine edit) pays nothing.
-    if (locallyChanged && baseline !== null && remoteSha !== undefined && baseline === remoteSha) {
-      try {
-        const remoteRawBody = await maybeDecryptFromPull(
-          await getBlobContent(token, owner, name, remoteSha),
-        )
-        if (isUnchangedModuloNormalization(content, remoteRawBody)) {
-          locallyChanged = false
-        }
-      } catch {
-        // Network blip / GC'd blob — fall back to the SHA decision (treat as a
-        // possible edit). Worst case is one re-push, never data loss.
-      }
-    }
-
-    noteBlobPlan.push({ path, content: wireContent, note, localSha, remoteSha, locallyChanged })
-  }
-  // push-only-real-edits: SUPPRESS the upload for a note that is NOT locally
-  // changed AND already has a remote blob at this path. We emit NO tree entry
-  // for it, so the base tree's existing (user's original, possibly
-  // non-canonical) blob is preserved untouched — zero rewrite. We STILL push
-  // when the note is locally changed (a real edit) OR has no remote blob yet
-  // (`remoteSha === undefined`: a brand-new note, or a note moved to a new path
-  // — the move's old-path deletion is handled by the deletion loop in step 4).
-  const noteBlobsSuppressed = noteBlobPlan.filter(p => !p.locallyChanged && p.remoteSha !== undefined && p.remoteSha !== p.localSha)
-  const suppressedNoteIds = new Set(noteBlobsSuppressed.map(p => p.note.id))
-  const noteBlobsToUpload = noteBlobPlan.filter(p => !suppressedNoteIds.has(p.note.id) && p.remoteSha !== p.localSha && !uploadedShas.has(p.localSha))
-  const noteBlobsCached   = noteBlobPlan.filter(p => !suppressedNoteIds.has(p.note.id) && p.remoteSha !== p.localSha &&  uploadedShas.has(p.localSha))
-
-  let blobsUploaded = 0
-  let blobsSkipped = noteBlobsCached.length
-  // Use a single running `total` that we refine after the attachment
-  // pre-pass below. For now: just notes.
-  let blobsTotal = noteBlobsToUpload.length
-
-  const emitBlobProgress = () => {
-    onProgress?.({ phase: 'uploading-blobs', uploaded: blobsUploaded, total: blobsTotal, skipped: blobsSkipped })
-  }
-
-  // Apply the cached-skip entries first — no network, just emit tree entries
-  // and bookkeeping.
-  for (const plan of noteBlobsCached) {
-    entries.push({ path: plan.path, mode: '100644', type: 'blob', sha: plan.localSha })
-    if (plan.remoteSha) updated++; else created++
-  }
-  // Pure-skip entries (remote has this SHA): no entry, no upload, but the
-  // note's path metadata may still need an update. The lastPushedSnapshot
-  // gets the PLAINTEXT body (gutter compares against unencrypted text).
-  //
-  // push-only-real-edits: a SUPPRESSED note (unchanged but its canonical wire
-  // SHA differs from the non-canonical remote blob) is left ENTIRELY untouched:
-  // no tree entry (handled by the filters above) AND no pathUpdate. Emitting a
-  // pathUpdate here would rewrite gitLastPushedSha to `finalSha` (the wire SHA),
-  // overwriting the canonical baseline syncApply pinned — which would make the
-  // NEXT pull misclassify the note (localChanged would flip on every sync). The
-  // note did not change, so we leave gitPath / gitLastPushedSha / gitRemoteBaseSha
-  // exactly as they are. We still take a gutter snapshot (body is unchanged, so
-  // the plaintext is correct).
-  for (const plan of noteBlobPlan) {
-    const suppressed = suppressedNoteIds.has(plan.note.id)
-    const skipped = plan.remoteSha === plan.localSha
-    if (!suppressed && skipped) {
-      // After a push the pushed blob IS the remote file, so the local baseline
-      // and the remote merge base coincide — set both to finalSha.
-      const finalSha = plan.remoteSha!
-      if (plan.note.gitPath !== plan.path || plan.note.gitLastPushedSha !== finalSha || plan.note.gitRemoteBaseSha !== finalSha) {
-        pathUpdates.push({ noteId: plan.note.id, gitPath: plan.path, gitLastPushedSha: finalSha, gitRemoteBaseSha: finalSha })
-      }
-    }
-    // Gutter snapshot uses the plaintext, not the wire form — look it up
-    // from `desired` rather than `plan.content` (which is the wire form).
-    const desiredEntry = desired.get(plan.path)
-    if (desiredEntry) {
-      lastPushedToSnapshot.push({ noteId: plan.note.id, content: desiredEntry.content })
-    }
-  }
-
-  // Upload the genuinely-changed blobs.
-  if (blobsTotal > 0) emitBlobProgress()
-  for (const plan of noteBlobsToUpload) {
-    const finalSha = await createBlob(token, owner, name, plan.content)
-    // Cache by LOCAL SHA — that's what the next iteration computes from
-    // the same content. (In production localSha === serverSha because
-    // both follow git's content-addressing, but the local key is what
-    // gates the next cache lookup.)
-    uploadedShas.add(plan.localSha)
-    entries.push({ path: plan.path, mode: '100644', type: 'blob', sha: finalSha })
-    if (plan.remoteSha) updated++; else created++
-    if (plan.note.gitPath !== plan.path || plan.note.gitLastPushedSha !== finalSha || plan.note.gitRemoteBaseSha !== finalSha) {
-      // Pushed blob == remote file → both SHAs coincide.
-      pathUpdates.push({ noteId: plan.note.id, gitPath: plan.path, gitLastPushedSha: finalSha, gitRemoteBaseSha: finalSha })
-    }
-    blobsUploaded++
-    emitBlobProgress()
-  }
-
-  // Fire-and-forget the per-note snapshot writes. The gutter will pick
-  // them up on the next render — we don't await because a slow IDB
-  // flush shouldn't block the push completing.
-  void (async () => {
-    const { setLastPushedContent } = await import('./lastPushedContent')
-    for (const { noteId, content } of lastPushedToSnapshot) {
-      try { await setLastPushedContent(noteId, content) } catch { /* silent — gutter is best-effort */ }
-    }
-  })()
-
-  // 3b. Local attachments → binary blob entries. Push uploads any local
-  // attachment whose SHA differs from the remote. Files only present
-  // locally get created remotely; files present in both get updated when
-  // their content drifts. Same upload-cache + progress treatment as notes.
-  const localAttachmentPaths = await listAttachmentPaths()
-  interface AttachmentPlan { path: string; localSha: string; remoteSha: string | undefined }
-  const attachmentPlan: AttachmentPlan[] = []
-  for (const path of localAttachmentPaths) {
-    if (pushMatcher.isIgnored(path)) continue
-    const localSha = await getAttachmentGitSha(path)
-    if (!localSha) continue
-    const remoteSha = remoteTree.get(path)
-    attachmentPlan.push({ path, localSha, remoteSha })
-  }
-  const attachmentsToUpload = attachmentPlan.filter(p => p.remoteSha !== p.localSha && !uploadedShas.has(p.localSha))
-  const attachmentsCached   = attachmentPlan.filter(p => p.remoteSha !== p.localSha &&  uploadedShas.has(p.localSha))
-  blobsTotal += attachmentsToUpload.length
-  blobsSkipped += attachmentsCached.length
-  if (blobsTotal > 0 || blobsSkipped > 0) emitBlobProgress()
-
-  for (const plan of attachmentsCached) {
-    entries.push({ path: plan.path, mode: '100644', type: 'blob', sha: plan.localSha })
-    if (plan.remoteSha) updated++; else created++
-  }
-  for (const plan of attachmentsToUpload) {
-    const blob = await getAttachmentBlob(plan.path)
-    if (!blob) continue
-    const uploadedSha = await createBlobBinary(token, owner, name, blob)
-    // See the note loop above: cache the LOCAL sha for the next-pass
-    // lookup, which uses local-side hashing.
-    uploadedShas.add(plan.localSha)
-    entries.push({ path: plan.path, mode: '100644', type: 'blob', sha: uploadedSha })
-    if (plan.remoteSha) updated++; else created++
-    blobsUploaded++
-    emitBlobProgress()
-  }
-
-  // 3c. Apply attachment tombstones — paths the user explicitly deleted
-  // locally need to be removed from the remote tree, otherwise pull would
-  // re-download them every cycle (the orphan-comes-back bug). We only
-  // delete entries that actually exist remotely; stale tombstones (file
-  // already gone remotely) get cleared too.
-  const tombstones = await getAttachmentTombstones()
-  const consumedTombstones: string[] = []
-  for (const path of tombstones) {
-    if (remoteTree.has(path)) {
-      entries.push({ path, mode: '100644', type: 'blob', sha: null })
-      deleted++
-    }
-    consumedTombstones.push(path)
-  }
-
-  // 3d-pre. Vault `.gitignore` editor draft (gi9n Settings UI). Push
-  // a tree entry for `.gitignore` only when the caller provided a
-  // draft AND it differs from the current remote content. Skipping
-  // the no-change case keeps idle syncs commit-free.
-  let vaultGitignorePushed = false
-  if (vaultGitignoreDraft != null && vaultGitignoreDraft !== pushRemoteRaw) {
-    const blobSha = await createBlob(token, owner, name, vaultGitignoreDraft)
-    entries.push({ path: GITIGNORE_PATH, mode: '100644', type: 'blob', sha: blobSha })
-    if (remoteGitignoreSha) updated++; else created++
-    vaultGitignorePushed = true
-  }
-
-  // 3d. Vault settings file (vs8x). Include it in the push tree when
-  // either (a) the local hash differs from the last pushed hash (the
-  // user changed something), or (b) the file is missing remotely (first
-  // push for this vault). Skip altogether if the caller didn't pass a
-  // vaultSettings bundle (settings sync disabled).
-  let vaultSettingsHashPushed: string | undefined
-  if (vaultSettings) {
-    const { path, content, contentHash, lastPushedHash } = vaultSettings
-    const remoteHasFile = remoteTree.has(path)
-    const localChanged = contentHash !== lastPushedHash
-    if (localChanged || !remoteHasFile) {
-      const blobSha = await createBlob(token, owner, name, content)
-      entries.push({ path, mode: '100644', type: 'blob', sha: blobSha })
-      if (remoteHasFile) updated++; else created++
-      vaultSettingsHashPushed = contentHash
-    } else {
-      // Already up to date on remote — surface the hash so the caller
-      // doesn't need to re-derive it.
-      vaultSettingsHashPushed = contentHash
-    }
-  }
-
-  // 4. Handle deletions: notes that USED to have a gitPath but no longer
-  // resolve there (rename or moved folder) and notes that are now in trash.
-  const desiredPaths = new Set(desired.keys())
-  const seenGitPaths = new Set<string>()
-
-  // rename-not-delete HARD PUSH-SIDE SAFETY NET (the critical data-loss
-  // preventer). Before we emit ANY `sha:null` delete, build the set of remote
-  // paths that a LIVE (non-deleted) note still represents. A remote file at
-  // such a path MUST NOT be deleted, even if some upstream classification was
-  // wrong (e.g. a rename misread as a delete that soft-deleted the note).
-  //
-  // A path is protected when EITHER:
-  //   (a) it is in `desired` — an active note's CURRENT computed path maps to
-  //       it (the note literally lives there now), OR
-  //   (b) a live note's serialized content hash equals the remote blob SHA at
-  //       that path — the file's content IS a live note, just under a name
-  //       whose form no longer matches the note's stored path (the dash↔space
-  //       rename case). Deleting it would destroy the user's real note.
-  // We hash each live note ONCE here (and only for paths present in the remote
-  // tree, so we never pay for paths we wouldn't delete anyway).
-  const protectedRemotePaths = new Set<string>(desiredPaths)
-  {
-    // Map remote path → its blob SHA, but only for paths some live note could
-    // be defending. We need a SHA→paths index to test content equality.
-    const livePlainShaByNote = new Map<string, string>()
-    for (const note of activeNotes) {
-      const sha = await gitBlobSha(serializeNote(note))
-      livePlainShaByNote.set(note.id, sha)
-    }
-    const liveShaSet = new Set(livePlainShaByNote.values())
-    // Also count any baseline SHA a live note last pushed — a note whose body
-    // hasn't been re-serialized identically (e.g. non-canonical remote) is
-    // still defended by the SHA it was pushed as.
-    for (const note of activeNotes) {
-      if (note.gitLastPushedSha) liveShaSet.add(note.gitLastPushedSha)
-    }
-    for (const [path, remoteSha] of remoteTree) {
-      if (protectedRemotePaths.has(path)) continue
-      if (liveShaSet.has(remoteSha)) protectedRemotePaths.add(path)
-    }
-  }
-
-  for (const note of activeNotes) {
-    if (note.gitPath) seenGitPaths.add(note.gitPath)
-    if (note.gitPath && !desiredPaths.has(note.gitPath) && remoteTree.has(note.gitPath)) {
-      // Safety net: never delete a path a live note still represents (by
-      // content), even though this note's CURRENT path moved away from it.
-      if (protectedRemotePaths.has(note.gitPath)) continue
-      entries.push({ path: note.gitPath, mode: '100644', type: 'blob', sha: null })
-      deleted++
-    }
-  }
-  for (const note of notes) {
-    if (note.isDeleted && note.gitPath && remoteTree.has(note.gitPath)) {
-      // Only delete if no active note has already moved into that path AND no
-      // live note's content maps to it. The protectedRemotePaths check is the
-      // hard guard: a soft-deleted note's gitPath that an active note still
-      // represents (by current path OR by content hash) must survive — the
-      // delete classification was a rename misread, not a real deletion.
-      if (
-        !desired.has(note.gitPath) &&
-        !seenGitPaths.has(note.gitPath) &&
-        !protectedRemotePaths.has(note.gitPath)
-      ) {
-        entries.push({ path: note.gitPath, mode: '100644', type: 'blob', sha: null })
-        deleted++
-      }
-      pathUpdates.push({ noteId: note.id, gitPath: null, gitLastPushedSha: null, gitRemoteBaseSha: null })
-    }
-  }
-
-  if (entries.length === 0) {
-    // Even with no tree changes, clear stale tombstones (files already gone
-    // remotely) so they don't re-attempt every sync.
-    if (consumedTombstones.length > 0) await clearAttachmentTombstones(consumedTombstones)
-    return {
-      result: { unchanged: true, created: 0, updated: 0, deleted: 0, commitSha: parentCommitSha, commitUrl: null },
-      pathUpdates,
-      vaultSettingsHashPushed,
-      vaultGitignorePushed,
-    }
-  }
-
-  // 5. Create new tree → commit → fast-forward branch. Each step gets
-  // its own progress event so the UI (and any error) can pinpoint where
-  // a failure happened.
-  onProgress?.({ phase: 'creating-tree' })
-  const newTreeSha = await createTree(token, owner, name, baseTreeSha, entries)
-  // Some "changed" entries can resolve to a blob byte-identical to what the
-  // base tree already holds (e.g. a freshly-cloned note that round-trips to
-  // the same bytes on the first sync). GitHub then returns a tree equal to the
-  // base, so committing it would create an EMPTY "Sync from Noteser (1 change)"
-  // commit — "No files changed" — cluttering the history on every initial sync
-  // (and, with discard-on-switch re-cloning, on every repo switch). Skip the
-  // commit entirely when the tree did not actually change.
-  if (newTreeSha === baseTreeSha) {
-    if (consumedTombstones.length > 0) await clearAttachmentTombstones(consumedTombstones)
-    uploadedShas.clear()
-    onProgress?.({ phase: 'done' })
-    return {
-      result: { unchanged: true, created: 0, updated: 0, deleted: 0, commitSha: parentCommitSha, commitUrl: null },
-      pathUpdates,
-      vaultSettingsHashPushed,
-      vaultGitignorePushed,
-    }
-  }
-  const total = created + updated + deleted
-  const autoMessage = `Sync from Noteser (${total} change${total === 1 ? '' : 's'})`
-  const message = commitMessage && commitMessage.length > 0 ? commitMessage : autoMessage
-  onProgress?.({ phase: 'creating-commit' })
-  const { sha: commitSha, html_url } = await createCommit(token, owner, name, message, newTreeSha, parentCommitSha)
-  onProgress?.({ phase: 'updating-ref' })
-  await updateBranchRef(token, owner, name, branch, commitSha)
-
-  // Push succeeded — drop tombstones whose deletes are now in the commit
-  // AND clear the upload cache for this repo. The next push will start
-  // from scratch (which is fine — remote tree will be consulted again).
-  if (consumedTombstones.length > 0) await clearAttachmentTombstones(consumedTombstones)
-  uploadedShas.clear()
-  onProgress?.({ phase: 'done' })
-
-  return {
-    result: { unchanged: false, created, updated, deleted, commitSha, commitUrl: html_url },
-    pathUpdates,
-    vaultSettingsHashPushed,
-    vaultGitignorePushed,
-  }
-}
+// External callers (hooks, components, tests) keep importing from
+// `@/utils/githubSync` so the split is transparent. New code that lives
+// inside the sync pipeline should import directly from the submodule it
+// belongs in.
+
+export {
+  // Path computation
+  notePath,
+  pushPath,
+  // MIME helpers (also used by tests)
+  guessMimeFromPath,
+  // Note serialization + parser
+  serializeNote,
+  normalizeForPush,
+  isUnchangedModuloNormalization,
+  parseNote,
+  // Public sync result shape
+  type ParsedNote,
+  type SyncResult,
+} from './githubSync/internal'
+
+export {
+  pullFromGitHub,
+  pullFromZipball,
+  takeZipballAttachmentBytes,
+} from './githubSync/syncPull'
+
+export {
+  syncToGitHub,
+  _resetUploadedShaCache,
+  type PushProgress,
+  type SyncInput,
+  type SyncOutcome,
+  type GitPathUpdate,
+} from './githubSync/syncPush'
+
+export {
+  type PullClassification,
+  type PullOutcome,
+} from './githubSync/syncClassify'
